hp 관련 변수들 int형에서 float형으로 수정

+++ Assets/01.Script/Base/User.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 [System.Serializable]
 public class User 
 {
-    public int maxHp;
-    public int hp;
+    public float maxHp;
+    public float hp;
     public float experience;
     public int level;
     public List<Transform> researcherFemaleSpawnPos;
